refactor(domain): move port interfaces under src/domain/ports/
Move the port interfaces (QueueStore, AgentRunner, GitHubClient and its
associated types, WorkspaceManager / handles / push options, LogStore,
ProcessRunner, InstructionLoader) from src/infra/<area>/<port>.ts into
src/domain/ports/<port>.ts. Implementations stay in src/infra/ and now
implement against the domain port. AgentRunInput/AgentRunResult remain
in src/domain/agent.ts because they are domain value types, not ports.

Side effect: services/ and daemon/ no longer import port interfaces
from infra/. The dependency edge that crosses layers now flows
domain ← services ← infra (composition) instead of services → infra
(interface lookup).

Resolves #14
