+++ src/domain/ports/instruction-loader.ts
@@ -1,0 +1,5 @@
+import type { InstructionDefinition } from "../instruction.js";
+
+export interface InstructionLoader {
+  loadById(instructionId: string): Promise<InstructionDefinition>;
+}
