+++ src/domain/ports/log-store.ts
