+++ src/domain/ports/process-runner.ts
@@ -1,0 +1,18 @@
+export interface RunProcessInput {
+  command: string;
+  args?: string[];
+  cwd?: string;
+  env?: NodeJS.ProcessEnv;
+  stdin?: string;
+  timeoutMs?: number;
+}
+
+export interface RunProcessResult {
+  exitCode: number;
+  stdout: string;
+  stderr: string;
+}
+
+export interface ProcessRunner {
+  run(input: RunProcessInput): Promise<RunProcessResult>;
+}
